Merge in SAT/pancake from feature/TAG-5193-simd-dp-chaining to develop

* commit '0f88b1bd6cea7f8a646b3b24a85401bf293c877e': (49 commits)
  Resolved an ASAN issue in src/pancake/SeedDBWriter.cpp.
  Updated bamboo_build.sh, Makefile and scripts/ci/configure_debug_sanitize_fallback.sh.
  Updated bamboo_build.sh and scripts/ci/configure.sh.
  Added a trailing comma to meson.build.
  Brackets and aesthetics in src/pancake/DPChain.cpp.
  Removed special cases in tests/src/test_MapperBatchGPU.cpp.
  Renamed NUM_REGISTERS to NUM_ELEMENTS and REGISTER_SIZE to ELEMENT_SIZE.
  Cleanup in src/pancake/DPChainSimd.cpp.
  Added a cassert include to src/pancake/DPChain.cpp.
  Attempt to reduce the number of instructions for computing the logPart in the inner loop in src/pancake/DPChainSimd.cpp.
  Instead of computing jVec in the inner loop, computing them during initialization.
  Passing Meson options in scripts/ci/configure_debug_sanitize_fallback.sh.
  Added the env variables to the conf-debug2 tool in Makefile.
  Updated the SeedOccurrenceThresholds_LoadFromFile test in tests/src/test_MapperCLR.cpp. Now, results should be identical for SIMD and SISD chaining.
  Using the std::tie to compare values in the max vector, in src/pancake/DPChainSimd.cpp. Added some comments and renamed a couple of variables.
  Updated tests in tests/src/test_MapperCLR.cpp. A couple of tests improved after I fixed the tie resolving when selecting the best score.
  Two bugfixes: (1) the 'c' conditional variable now includes a check if the target ID and strand are the same. This handles the first and the last vector where edge cases may arise. However, this is not optimal, because other vectors in the loop don't require this check, so these are wasted cycles. (2) The linear pass over the vector with maximum scores also needed to check that the potential better predecessor index is larger than the current best - otherwise some overlapping seed hits may be lost, because elements in this vector can be out of order (e.g. the first element could come from a later vector than the second element).
  Updated an edge case test in tests/src/test_DPChainSimd.cpp, some values were wrong.
  Bugfix: the predecessor cannot be chosen if the predecessor index is >= i. This limits the impact of the last vector. It may have a downside: the out-of-bounds elements of the last vector may shaddow some valid predecessors to 'i'. However, this is faster, and the edge case would be very rare (if possible) and not cause logical issues.
  Fixed a low-complexity test in tests/src/test_DPChainSimd.cpp. Because now the closest seed hit to 'i' is preferred, the expected output of one test needed to be updated.
  ...

+++ include/pacbio/pancake/SeedHit.h
@@ -121,8 +121,9 @@ public:
 inline std::ostream& operator<<(std::ostream& os, const SeedHit& a)
 {
     os << "tid = " << a.targetId << ", trev = " << a.targetRev << ", tpos = " << a.targetPos
-       << ", qpos = " << a.queryPos << ", tspan = " << a.targetSpan << ", qspan = " << a.querySpan
-       << ", flags = " << a.flags << ", diag = " << (a.targetPos - a.queryPos);
+       << ", qpos = " << a.queryPos << ", tspan = " << static_cast<int32_t>(a.targetSpan)
+       << ", qspan = " << static_cast<int32_t>(a.querySpan) << ", flags = " << a.flags
+       << ", diag = " << (a.targetPos - a.queryPos);
     return os;
 }
